Add basic test for tracing garbage collection
The test builds two random b-trees, and then uses one of
them as a gc root, and including the other in the lazy
sequence of "all keys" to be dealt with by the collector.
After running a gc, we assert that the `deleted-fn` has
been invoked against all the "dead" nodes.